extract to get.utils.ts
