+++ carbonmark-api/src/routes/purchases/[id]/get.utils.ts
@@ -1,0 +1,23 @@
+import { isHexString } from "ethers-v6";
+import { Purchase as MumbaiPurchase } from "../../../graphql/marketplaceMumbai.types";
+import { NetworkParam } from "../../../models/NetworkParam.model";
+import { Purchase } from "../../../models/Purchase.model";
+import { gqlSdk } from "../../../utils/gqlSdk";
+
+/** Purchase ids are a txn hash */
+export const isValidPurchaseId = (id?: string | null) => {
+  if (!id) return false;
+  return id.length === 66 && isHexString(id);
+};
+
+export const getPurchaseById = async (params: {
+  id: string | null;
+  network?: NetworkParam;
+}): Promise<Purchase | MumbaiPurchase | null> => {
+  const graph =
+    params.network === "mumbai" ? gqlSdk.marketplaceMumbai : gqlSdk.marketplace;
+  const response = await graph.getPurchasesById({
+    id: params.id,
+  });
+  return response.purchases?.at(0) || null;
+};
